+++ CHANGELOG.md
@@ -16,10 +16,12 @@ Increasing the minimal supported Rust version will always be coupled at least wi
 * Added a `#[dsl::auto_type]` attribute macro, allowing to infer type of query fragment functions
 * Added the same type inference on `Selectable` derives, which allows skipping specifying `select_expression_type` most of the time, in turn enabling most queries to be written using just a `Selectable` derive.
 * Added an optional `#[diesel(skip_insertion)]` field attribute to the `Insertable` derive macro, allowing fields which map to generated columns to be skipped during insertion.
+* Support for connection instrumentation. This allows to inspect any query run by your application
+* Logging in diesel-cli
 
 ### Changed
 
-* The minimal officially supported rustc version is now 1.66.0
+* The minimal officially supported rustc version is now 1.70.0
 
 ## [2.1.0] 2023-05-26